account now belongs to one user

+++ api/src/services/account.js
@@ -13,10 +13,11 @@ export default class AccountService {
     number,
     name,
     description,
+    userId
   }) {
     try {
       const account = await this.accountModel.create(
-        { number, name, description });
+        { number, name, description, userId });
       return account;
     } catch (err) {
       this.logger.error(err);
@@ -24,8 +25,8 @@ export default class AccountService {
     }
   }
 
-  async findAll() {
-    const accounts = await this.accountModel.findAll({  });
+  async findAll(userId) {
+    const accounts = await this.accountModel.findAll({ where: { userId } });
     return accounts.map((a) => {
       return ({
         id: a.id,
@@ -39,8 +40,8 @@ export default class AccountService {
     });
   }
 
-  async findById(id) {
-    const account = await this.accountModel.findOne({ where: { id } });
+  async findById(id, userId) {
+    const account = await this.accountModel.findOne({ where: { id, userId } });
     if (!account) {
       return null;
     }
@@ -55,16 +56,16 @@ export default class AccountService {
     });
   }
 
-  async updateById(id, values) {
-    const affectedRows = await this.accountModel.update(values, { where: { id } });
+  async updateById(id, userId, values) {
+    const affectedRows = await this.accountModel.update(values, { where: { id, userId } });
     if (affectedRows === 0) {
       return null;
     }
-    return this.findById(id);
+    return this.findById(id, userId);
   }
 
-  async deleteById(id) {
-    const affectedRows = await this.accountModel.destroy({ where: { id } });
+  async deleteById(id, userId) {
+    const affectedRows = await this.accountModel.destroy({ where: { id, userId } });
     if (affectedRows === 0) {
       throw new Error('Account does not exist');
     }
